+++ packages/material-hooks/src/basic/solid.frag.ts
