+++ packages/material-hooks/src/basic/solid.vert.ts
