+++ packages/material-hooks/src/lambert/index.ts
@@ -12,35 +12,33 @@ import solidVert from './solid.vert'
 import solidFrag from './solid.frag'
 import texturedVert from './textured.vert'
 import texturedFrag from './textured.frag'
-import attenuatedVert from './attenuated.vert'
-import attenuatedFrag from './attenuated.frag'
 
 const defaultKd = [1.0, 1.0, 1.0]
-const defaultKs = [0.6, 0.6, 0.6]
-const defaultNs = 500
 const defaultKa = [1.0, 1.0, 1.0]
 const defaultNa = 0
 
-export function usePhongSolid(kd, na, ns, ka, ks) {
+export function useLambertSolid(kd: number[], na: number, ka: number[]) {
   const gl = useWebGLContext()
   const [vert, frag] = usePointLightCount(solidVert, solidFrag)
   const program = useProgram(gl, vert, frag)
 
   usePointLightUniforms(gl, program)
 
-  useUniform1f(gl, program, 'uNs', ns || defaultNs)
   useUniform1f(gl, program, 'uNa', na || defaultNa)
+  useUniform3fv(gl, program, 'uKa', ka || defaultKa)
 
   useUniform3fv(gl, program, 'uKd', kd || defaultKd)
-  useUniform3fv(gl, program, 'uKs', ks || defaultKs)
-  useUniform3fv(gl, program, 'uKa', ka || defaultKa)
 
   return program
 }
 
 const uVScale = [1.0, 1.0]
 
-export function usePhongTextured(mapKd, na, ns, ka, ks) {
+export function useLambertTextured(
+  mapKd: WebGLTexture,
+  na: number,
+  ka: number[],
+) {
   const gl = useWebGLContext()
   const [vert, frag] = usePointLightCount(texturedVert, texturedFrag)
   const program = useProgram(gl, vert, frag)
@@ -49,24 +47,10 @@ export function usePhongTextured(mapKd, na, ns, ka, ks) {
 
   useUniformSampler2d(gl, program, 'mapKd', mapKd)
 
-  useUniform1f(gl, program, 'uNs', ns || defaultNs)
-  useUniform1f(gl, program, 'uNa', na || defaultNa)
-
   useUniform2fv(gl, program, 'uVScale', uVScale)
 
-  useUniform3fv(gl, program, 'uKs', ks || defaultKs)
+  useUniform1f(gl, program, 'uNa', na || defaultNa)
   useUniform3fv(gl, program, 'uKa', ka || defaultKa)
-
-  return program
-}
-
-export function usePhongAttenuated(lightPosition, mapKd) {
-  const gl = useWebGLContext()
-  const program = useProgram(gl, attenuatedVert, attenuatedFrag)
-
-  useUniform3fv(gl, program, 'uLightPosition', lightPosition)
-
-  useUniformSampler2d(gl, program, 'texDiff', mapKd)
 
   return program
 }
