+++ packages/material-hooks/src/phong/attenuated.frag.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 import glsl from 'glslify'
 
 export default glsl`
