+++ packages/material-hooks/src/phong/textured.frag.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 import glsl from 'glslify'
 
 export default glsl`
