+++ packages/material-hooks/src/phong/textured.vert.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 import glsl from 'glslify'
 
 export default glsl`
@@ -6,42 +7,37 @@ export default glsl`
   uniform mat4 viewMatrix;
   uniform mat4 modelMatrix;
   uniform mat4 projectionMatrix;
-
+  
   const int NUM_POINT_LIGHTS = <<NUM_POINT_LIGHTS>>;
-  uniform vec3 pointLd[NUM_POINT_LIGHTS];
   uniform vec3 pointLp[NUM_POINT_LIGHTS];
-
-  uniform vec3 uKd;
-  uniform vec3 uKa;
-  uniform float uNa;
-
+  
   attribute vec3 position;
   attribute vec3 normal;
+  attribute vec2 uv;
 
-  varying vec3 vColor;
+  varying vec3 vNormal;
+  varying vec2 vUv;
+  varying vec3 vEye;
+  varying vec3 vRay[NUM_POINT_LIGHTS];
 
   #pragma glslify: transpose = require('glsl-transpose')
   #pragma glslify: inverse = require('glsl-inverse')
 
-  void main(void) {
+  void main() {
     mat4 modelViewMatrix = viewMatrix * modelMatrix;
     vec4 viewModelPosition = modelViewMatrix * vec4(position, 1.0);
   
     mat3 normalMatrix = transpose(inverse(mat3(modelViewMatrix)));
-    vec3 N = normalize(normalMatrix * normal);
+    vNormal = vec3(normalMatrix * normal);
 
-    vec3 color = vec3(0.0);
-    vec3 light = vec3(0.0);
-    vec4 lightPosition = vec4(0.0);
-    vec3 ambient = uKd * uKa * uNa;
-
-    for (int i = 0; i < NUM_POINT_LIGHTS; i++) {
-      lightPosition = viewMatrix * vec4(pointLp[i], 1.0);
-      light = -normalize(viewModelPosition.xyz - lightPosition.xyz);
-      color += (pointLd[i] * uKd * clamp(dot(N, light), 0.0, 1.0));
+    for(int i = 0; i < NUM_POINT_LIGHTS; i++) {
+      vec4 lightPosition = viewMatrix * vec4(pointLp[i], 1.0);
+      vRay[i] = viewModelPosition.xyz - lightPosition.xyz;
     }
 
-    vColor = color + ambient;
+    vUv = uv;
+    vEye = -vec3(viewModelPosition.xyz);
+
 
     gl_Position = projectionMatrix * viewModelPosition;
   }
