+++ packages/material-hooks/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../base.tsconfig.json",
+  "include": ["./src/**/*"]
+}
